libbst: structurize
