+++ library/numeric/binary_trees/nodes/bst_structure.h
@@ -1,0 +1,15 @@
+#ifndef SPBU_2020_C_NUMERIC_NODES_BST_STRUCTURE_H
+#define SPBU_2020_C_NUMERIC_NODES_BST_STRUCTURE_H
+#include "../bst_node.h"
+
+#ifndef SPBU_2020_C_NUMERIC_NODES_AVL_STRUCTURE_H
+
+struct BinaryTreeNode {
+    int value;
+    BinaryTreeNode* leftChild;
+    BinaryTreeNode* rightChild;
+};
+
+#endif /* SPBU_2020_C_NUMERIC_NODES_AVL_STRUCTURE_H */
+
+#endif /* SPBU_2020_C_NUMERIC_NODES_BST_STRUCTURE_H */
